+++ lib/flutter_kurdish.dart
@@ -1,0 +1,5 @@
+library flutter_kurdish;
+
+export 'package:flutter_kurdish/kurdish_material_localization_delegate.dart';
+export 'package:flutter_kurdish/kurdish_widget_localization_delegate.dart';
+export 'package:flutter_kurdish/kurdish_cupertino_localization_delegate.dart';